+++ src/pages/Missions.vue
@@ -2,27 +2,27 @@
   <q-page class="q-pa-md">
     <q-breadcrumbs active-color="secondary" color="light">
       <q-breadcrumbs-el label="Home" to="/"/>
-      <q-breadcrumbs-el label="Tasks" to="/tasks"/>
+      <q-breadcrumbs-el label="Missions" to="/missions"/>
       <q-btn
         dense round icon="refresh" class="q-ml-lg" size="sm"
-        @click="refreshTasks"
+        @click="refreshMissions"
       />
     </q-breadcrumbs>
 
     <q-table
-      title="Tasks"
-      :data="allTasksList"
+      title="Missions"
+      :data="allMissionsList"
       :columns="columns"
       row-key="name"
     >
       <div slot="top-right" slot-scope="props" class="fit">
-        <task-new-button v-on:created="taskCreated"/>
+        <mission-new-button v-on:created="missionCreated"/>
       </div>
 
       <q-td slot="body-cell-name" slot-scope="props" :props="props"
             style="width:5px"
       >
-        <router-link :to="`/tasks/${encodeURIComponent(props.row.taskId)}`">
+        <router-link :to="`/missions/${encodeURIComponent(props.row.missionId)}`">
           {{props.row.name}}
         </router-link>
       </q-td>
@@ -32,19 +32,19 @@
 </template>
 
 <script>
-  import TaskNewButton from 'components/task-new-button'
-  import allTasksList from '../graphql/tasks.gql'
+  import MissionNewButton from 'components/mission-new-button'
+  import allMissionsList from '../graphql/missions.gql'
 
   const debug = false
 
   export default {
     components: {
-      TaskNewButton
+      MissionNewButton
     },
 
     data() {
       return {
-        allTasksList: [],
+        allMissionsList: [],
         columns: [
           {
             field: 'name',
@@ -65,26 +65,26 @@
     },
 
     apollo: {
-      allTasksList,
+      allMissionsList,
     },
 
     mounted() {
-      this.refreshTasks()
+      this.refreshMissions()
     },
 
     methods: {
-      refreshTasks() {
-        this.$apollo.queries.allTasksList.refetch()
+      refreshMissions() {
+        this.$apollo.queries.allMissionsList.refetch()
       },
 
-      taskCreated(data) {
-        this.refreshTasks()
+      missionCreated(data) {
+        this.refreshMissions()
       }
     },
 
     watch: {
-      allTasksList(val) {
-        if (debug) console.log('watch allTasksList=', val)
+      allMissionsList(val) {
+        if (debug) console.log('watch allMissionsList=', val)
       }
     }
   }